% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2po.R
\name{rmd2po}
\alias{rmd2po}
\alias{po2rmd}
\title{Create a poEdit file from an R Markdown or Quarto document, or translate such a document using the .po file}
\usage{
rmd2po(
  rmdfile,
  lang = "fr",
  podir = "po",
  mdpodir = getOption("mdpodir"),
  min.version = "2.0.0",
  verbose = FALSE,
  keep.tmpfile = FALSE
)

po2rmd(
  rmdfile,
  lang = "fr",
  podir = "po",
  mdpodir = getOption("mdpodir"),
  min.version = "2.0.0",
  verbose = FALSE,
  keep.tmpfile = FALSE
)
}
\arguments{
\item{rmdfile}{The path to the R Markdown or Quarto document to translate}

\item{lang}{The language to translate to, like \code{"fr"} for French, \code{"es"} for
Spanish, ... Also the subdirectory to the directory where the original file
is located where to place the translated .Rmd or .qmd file.}

\item{podir}{The subdirectory of \code{lang} where to place the .po file (by
default, it is \code{"po"})}

\item{mdpodir}{The directory that contains md2po and po2md programs (\code{NULL},
by default, if these programs are accessible directly from the command
line within the R process)}

\item{min.version}{The minimum version of md2po and po2md required (string
like "2.0.0")}

\item{verbose}{If \code{TRUE}, print more info about md2po or po2md and the
command that is executed}

\item{keep.tmpfile}{If \code{TRUE}, keep the modified .tmp file that is created
from the original .Rmd/.qmd to allow better handling of YAML header and R
chunks. \code{FALSE} by default, change it only for debugging purposes}
}
\value{
The path to the .po file (for \code{rmd2po()}) or to the translated
Rmd/qmd file (for \code{po2rmd()}) is returned. The .po file or the translated
Rmd/qmd file is created or updated on each call of the respective function.
}
\description{
Create a poEdit file from an R Markdown or Quarto document, or translate such a document using the .po file
}
\details{
This function internally uses md2po and po2md CLI programs that are
from the mdpo Python library. You have to install these before use and make
sure that md2po and po2md are accessible on the command line from within R,
or provide the absolute path where they are located in the \code{"mdpodir"} option
using something like \code{options(mdpodir = "/usr/bin")}.
md2po is not dealing well with YAML headers, chunk headers and list items in
the .Rmd/.qmd files. A special code '=====' is introduced in the strings to
"escape" them from a wrong processing by md2po. This code is removed in the
final translated .Rmd/.qmd file.
You should leave these "=====" codes in the translated string too in the .po
file.

\strong{A few advise when you translate the strings in poEdit:}
\itemize{
\item Use a correct syntax in your .Rmd/.qmd file. Indent by \strong{four} spaces when
required (do \strong{not} use two spaces to indent items inside lists, for
instance). Do not wrap paragraphs and place one empty line between each
block. That way, you will get an exact correspondence of line numbers between
the original and translated .Rmd/.qmd files. Bring correction in the layout of
the original file, if needed.
\item The string "#=====" tags a complex chunk header. You should not change it,
except, may be the content of \code{fig.cap="..."} that could be translated.
\item For chunks, it is easier to start from an identical copy using
Ctrl-B/Cmd-B. Most of the time, nothing or very little parts must be changed
(mostly comments, but also see next point here under). Take care of quotes ',
and " that could be changed, depending on the language used (e.g., French).
The French quotes are, of course, inappropriate in R code. If the change is
automatic in poEdit, Ctrl-Z/Cmd-Z undoes the change.
\item Where the .Rmd/.qmd refers to a document in the same directory or in a
subdirectory, remember that the translated vignette is located in a "\if{html}{\out{<lang>}}"
subdirectory. It means that the relative path must be prepended with "../" to
reflect its new location. For instance, for a dataset "flights14.csv",
\code{fread('fligth14.csv')} must be changed into \code{fread('../flights14.csv')} in
the translated version. Also for a figure "fig1.png" in, say the "plots"
subdirectory, \verb{![](plots/fig1.png)} must be changed into
\verb{![](../plots/fig1.png)}. Finally, documents referred in the YAML header must
also be changed accordingly. For a "style.css" file in the "css"
subdirectory, change \verb{css: [default, css/toc.css]} into
\verb{css: [default, ../css/toc.css]} in the YAML header of the translated .Rmd
file. This way, there is no need to duplicate files. Failure to do so will
result in an error during compilation of the .Rmd file. So, go back to
poEdit, correct the concerned item, and relaunch \code{po2rmd()} until compilation
runs flawlessly.
\item List items are surrounded by "=====" markers. This is to avoid po2md to
reinterpret them in a different way, and it is also to draw your attention on
pieces of text that may not be complete sentences. Keep the exact same tags
in the translated version. If list items are not separated by an empty line,
they are concatenated into a single string in poEdit, something like:
"=====1. First item===== =====2. Second item===== =====3. Third item====="
This is normal. Just keep the same tags and correct list items will be placed
in the final translated .Rmd/.qmd file by \code{rmd2po()}.
}
}
\examples{
# TODO: and example using a short vignette
}
