% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdpo-package.R
\docType{package}
\name{rmdpo-package}
\alias{rmdpo}
\alias{rmdpo-package}
\title{rmdpo: Translate R Markdown and Quarto documents using poEdit}
\description{
Extract each block of an R Markdown or Quarto document into a translatable string in .po format (editable in poEdit, or transferred into other translation tools). Once the strings are translated, build the translated version of the initial document in the same R Markdown or Quarto format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/rmdpo}
  \item \url{https://www.sciviews.org/rmdpo/}
  \item Report bugs at \url{https://github.com/SciViews/rmdpo/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

Authors:
\itemize{
  \item Chritian Wiat \email{w9204-r@yahoo.com}
}

}
\keyword{internal}
